Moving autocomplete and notebook extension from toolbox setup to engine template setup. Close #107
